更新 customize.sh

+++ module/customize.sh
@@ -95,10 +95,9 @@ else
     logowl "settings.conf already exists"
     logowl "settings.conf will NOT be overwritten"
 fi
-if [ -n "$VERIFY_DIR" ] && [ -d "$VERIFY_DIR" ] && [ "$VERIFY_DIR" != "/" ]; then
-    rm -rf "$VERIFY_DIR"
-fi
+[ "$VERIFY_DIR" != "/" ] && rm -rf "$VERIFY_DIR"
+logowl "Set permission"
 set_permission_recursive "$MODPATH" 0 0 0755 0644
 logowl "Welcome to use $MOD_NAME!"
-DESCRIPTION="[⏳Reboot to take effect. ✨Root: $ROOT_SOL_DETAIL] $MOD_INTRO"
-update_config_value "description" "$DESCRIPTION" "$MODPATH/module.prop" "true"
+DESCRIPTION="[⏳Reboot to take effect. ⚙️Root: $ROOT_SOL_DETAIL] $MOD_INTRO"
+update_config_value "description" "$DESCRIPTION" "$MODPATH/module.prop"
